refactor replace domains to data/model
