+++ lib/data/model/coordinate.dart
