+++ lib/data/model/enums/coordinate_type.dart
