+++ lib/data/model/enums/search_screen_state_type.dart
