+++ lib/data/model/enums/sight_type.dart
