+++ lib/ui/screen/sight_list_screen/sight_list_portrate_orientation.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../domain/sight.dart';
+import '../../../data/model/sight.dart';
 import '../../widgets/sight_cards/sight_card.dart';
 
 class SightListPortrateOrientation extends StatelessWidget {
