+++ lib/ui/screen/sight_search_screen/searched_sights_list_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/domain/searched_sight.dart';
+import 'package:places/data/model/searched_sight.dart';
 import 'package:places/ui/screen/sight_details_screen/sight_details_screen.dart';
 import 'package:places/ui/screen/sight_search_screen/sight_card_tile.dart';
